requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytesseract==0.3.10
pytz==2023.3.post1
sqlparse==0.4.4
